fix(balance): 怪物移速恢复为旧版整数值，Monster.speed改回int

+++ entities/monster.py
@@ -22,7 +22,7 @@ class Monster:
     x: float = 0.0
     y: float = 0.0
     cooldown_remaining: float = 0.0   # 当前剩余冷却（秒）
-    speed: float = MONSTER_SPEED
+    speed: int = MONSTER_SPEED
     alive: bool = True
     aggro: bool = False       # 是否已进入战斗状态
